pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cell-cli"
version = "0.1.0"
description = "A small footprint pseudo-node agent for massive agentic swarms"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "textual>=0.47.0",
    "litellm>=1.20.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pyyaml>=6.0"
]

[project.scripts]
cell = "cell.app:main"
